Consolidate ClientIpHeaderFilter tests
