+++ services/spring-cloud-gateway/src/test/java/unit/net/firedevops/firemud/filter/ClientIpHeaderFilterTest.java
@@ -1,9 +1,10 @@
 package net.firedevops.firemud.filter;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
 import java.net.InetSocketAddress;
+import java.util.concurrent.atomic.AtomicReference;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.http.server.reactive.MockServerHttpRequest;
 import org.springframework.mock.web.server.MockServerWebExchange;
@@ -12,55 +13,71 @@ import org.springframework.web.server.WebFilterChain;
 import reactor.core.publisher.Mono;
 
 class ClientIpHeaderFilterTest {
+
+  private final ClientIpHeaderFilter filter = new ClientIpHeaderFilter();
+
   @Test
   void addsClientIpHeader() {
-    ClientIpHeaderFilter filter = new ClientIpHeaderFilter();
     MockServerHttpRequest request =
         MockServerHttpRequest.get("/").remoteAddress(new InetSocketAddress("1.2.3.4", 0)).build();
-    MockServerWebExchange exchange = MockServerWebExchange.from(request);
-    java.util.concurrent.atomic.AtomicReference<ServerWebExchange> ref =
-        new java.util.concurrent.atomic.AtomicReference<>();
-    WebFilterChain chain =
-        e -> {
-          ref.set(e);
-          return Mono.empty();
-        };
-    filter.filter(exchange, chain).block();
-    assertEquals("1.2.3.4", ref.get().getRequest().getHeaders().getFirst("X-Client-IP"));
+    ServerWebExchange mutatedExchange = filterThroughChain(MockServerWebExchange.from(request));
+
+    assertThat(mutatedExchange.getRequest().getHeaders().getFirst("X-Client-IP"))
+        .isEqualTo("1.2.3.4");
   }
 
   @Test
   void noHeaderWhenRemoteAddressMissing() {
-    ClientIpHeaderFilter filter = new ClientIpHeaderFilter();
     MockServerHttpRequest request = MockServerHttpRequest.get("/").build();
-    MockServerWebExchange exchange = MockServerWebExchange.from(request);
-    java.util.concurrent.atomic.AtomicReference<ServerWebExchange> ref =
-        new java.util.concurrent.atomic.AtomicReference<>();
-    WebFilterChain chain =
-        e -> {
-          ref.set(e);
-          return Mono.empty();
-        };
-    filter.filter(exchange, chain).block();
-    assertNull(ref.get().getRequest().getHeaders().getFirst("X-Client-IP"));
+    ServerWebExchange mutatedExchange = filterThroughChain(MockServerWebExchange.from(request));
+
+    assertNull(mutatedExchange.getRequest().getHeaders().getFirst("X-Client-IP"));
   }
 
   @Test
   void noHeaderWhenAddressUnresolved() {
-    ClientIpHeaderFilter filter = new ClientIpHeaderFilter();
     MockServerHttpRequest request =
         MockServerHttpRequest.get("/")
             .remoteAddress(java.net.InetSocketAddress.createUnresolved("example.com", 0))
             .build();
-    MockServerWebExchange exchange = MockServerWebExchange.from(request);
-    java.util.concurrent.atomic.AtomicReference<ServerWebExchange> ref =
-        new java.util.concurrent.atomic.AtomicReference<>();
+    ServerWebExchange mutatedExchange = filterThroughChain(MockServerWebExchange.from(request));
+
+    assertNull(mutatedExchange.getRequest().getHeaders().getFirst("X-Client-IP"));
+  }
+
+  @Test
+  void preservesExistingClientIpHeader() {
+    MockServerHttpRequest request =
+        MockServerHttpRequest.get("http://localhost/api/session/test")
+            .header("X-Client-IP", "203.0.113.10")
+            .build();
+
+    ServerWebExchange mutatedExchange = filterThroughChain(MockServerWebExchange.from(request));
+
+    assertThat(mutatedExchange.getRequest().getHeaders().getFirst("X-Client-IP"))
+        .isEqualTo("203.0.113.10");
+  }
+
+  @Test
+  void derivesClientIpFromRemoteAddressWhenHeaderMissing() {
+    MockServerHttpRequest request =
+        MockServerHttpRequest.get("http://localhost/ws/game/test")
+            .remoteAddress(new InetSocketAddress("198.51.100.7", 443))
+            .build();
+    ServerWebExchange mutatedExchange = filterThroughChain(MockServerWebExchange.from(request));
+
+    assertThat(mutatedExchange.getRequest().getHeaders().getFirst("X-Client-IP"))
+        .isEqualTo("198.51.100.7");
+  }
+
+  private ServerWebExchange filterThroughChain(ServerWebExchange exchange) {
+    AtomicReference<ServerWebExchange> ref = new AtomicReference<>();
     WebFilterChain chain =
         e -> {
           ref.set(e);
           return Mono.empty();
         };
     filter.filter(exchange, chain).block();
-    assertNull(ref.get().getRequest().getHeaders().getFirst("X-Client-IP"));
+    return ref.get();
   }
 }
